[#24] refactor: 파사드를 위한 프로젝트 패키지 구조 변경
