+++ src/main/java/goorm/eagle7/stelligence/common/merge/handler/AmendmentMergeTemplate.java
@@ -3,7 +3,7 @@ package goorm.eagle7.stelligence.common.merge.handler;
 import org.springframework.stereotype.Component;
 
 import goorm.eagle7.stelligence.domain.amendment.model.Amendment;
-import goorm.eagle7.stelligence.domain.document.model.Document;
+import goorm.eagle7.stelligence.domain.document.content.model.Document;
 import goorm.eagle7.stelligence.domain.section.SectionRepository;
 import goorm.eagle7.stelligence.domain.section.model.Section;
 import lombok.RequiredArgsConstructor;
