+++ src/main/java/goorm/eagle7/stelligence/domain/document/content/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package goorm.eagle7.stelligence.domain.document.dto;
+package goorm.eagle7.stelligence.domain.document.content.dto;
 
 import goorm.eagle7.stelligence.domain.section.model.Heading;
 import lombok.AllArgsConstructor;
