+++ src/main/java/goorm/eagle7/stelligence/domain/section/model/Section.java
@@ -4,7 +4,7 @@ import static jakarta.persistence.FetchType.*;
 import static lombok.AccessLevel.*;
 
 import goorm.eagle7.stelligence.common.entity.BaseTimeEntity;
-import goorm.eagle7.stelligence.domain.document.model.Document;
+import goorm.eagle7.stelligence.domain.document.content.model.Document;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
