+++ src/test/java/goorm/eagle7/stelligence/common/merge/MergeServiceIntegratedTest.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import goorm.eagle7.stelligence.config.mockdata.WithMockData;
 import goorm.eagle7.stelligence.domain.contribute.model.Contribute;
-import goorm.eagle7.stelligence.domain.document.model.Document;
+import goorm.eagle7.stelligence.domain.document.content.model.Document;
 import goorm.eagle7.stelligence.domain.section.model.Section;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
